Move GetCmdOverrideAppProperty to utility public function
